finished business logic and some refactor

+++ Models/ReisefradragResult.cs
@@ -2,6 +2,6 @@ namespace reisefradragApi.Models
 {
     public class ReisefradragResult
     {
-        public int Reisefradrag { get; set; }
+        public double Reisefradrag { get; set; }
     }
 }
